[Search] Add sample search python impl.

+++ include/models.h
@@ -17,7 +17,7 @@ public:
 
     void config(int maxLen_ = -1, int numBeams_ = 1, int numBeamHypsToKeep_ = 1, float lenPenalty_ = 1.0,
             bool doEarlyStopping_ = false, int eosTokenId_ = -1, int padTokenId_ = -1, bool doSample_ = false,
-            float temperature_ = 1.0, float topK_ = 50, float topP_ = 1.0);
+            float temperature_ = 1.0, int topK_ = 50, float topP_ = 1.0);
 
     bool isDone();
